dom/media/webaudio/test/test_waveShaperNoCurve.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test WaveShaperNode with no curve</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="webaudio.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

var gTest = {
  length: 2048,
  numberOfChannels: 1,
  createGraph: function(context) {
    var source = context.createBufferSource();
    source.buffer = this.buffer;

    var shaper = context.createWaveShaper();
    is(shaper.curve, null, "The shaper curve must be null by default");

    source.connect(shaper);

    source.start(0);
    return shaper;
  },
  createExpectedBuffers: function(context) {
    var expectedBuffer = context.createBuffer(1, 2048, context.sampleRate);
    for (var i = 0; i < 2048; ++i) {
      expectedBuffer.getChannelData(0)[i] = Math.sin(440 * 2 * Math.PI * i / context.sampleRate);
    }
    this.buffer = expectedBuffer;
    return expectedBuffer;
  },
};

runTest();

</script>
</pre>
</body>
</html>
